fixed services and classes for batch and associate

+++ src/main/java/com/revature/backend/repository/BatchRepository.java
@@ -5,13 +5,12 @@ import java.util.List;
 import javax.transaction.Transactional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.revature.backend.model.Batch;
 
 //@Transactional
 @Repository
-public interface BatchRepository extends JpaRepository<Batch, Long>{
+public interface BatchRepository extends JpaRepository<Batch, Integer>{
 
 }
